requirements.txt
pandas==2.0.3
numpy==1.25.1
PyJWT==2.8.0
st-pages==0.4.3
streamlit==1.24.1
streamlit-option-menu==0.3.6
requests==2.28.2
pycryptodome==3.18.0